requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.80.0
frozenlist==1.3.1
h11==0.13.0
httptools==0.4.0
idna==3.3
ijson==2.6.1
loguru==0.6.0
meraki==1.22.1
multidict==6.0.2
oauthlib==3.2.0
pkg_resources==0.0.0
pydantic==1.9.2
pyprtg-api==0.0.8
pysnow==0.7.17
python-dotenv==0.20.0
python-magic==0.4.27
pytz==2019.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarl==1.8.1
